[pdfjs-images] - base for extracting images with pdfjs
