+++ server/src/processing/ImageDetectionModule/extractors/PdfJsImageExtractor.ts
@@ -1,0 +1,26 @@
+/**
+ * Copyright 2020 AXA Group Operations S.A.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { Document } from '../../../types/DocumentRepresentation';
+import logger from '../../../utils/Logger';
+import { ImageExtractor } from "./ImageExtractor";
+
+export class PdfJsImageExtractor extends ImageExtractor {
+  public async run(_doc: Document): Promise<void> {
+    logger.warn('PDFjs Image extractor not implemented');
+    return;
+  }
+}
